+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## v1.1.0 2017-05-29
+- Added support for using one interrupt GPIO pin for multiple instances of the PCF8574 class
+
 ## v1.0.2 2017-05-27
-- internal currentState gets now updated immediatly to avoid overlaps between different calls
+- Internal currentState gets now updated immediatly to avoid overlaps between different calls
 
 ## v1.0.1 2017-04-17
 - Small fixes in readme and examples